Allow docker containers to download plugins
It allows automatic installation of plugins for docker.

There can be multiple plugins, with one variable each. As long as the
environment variable starts with "CPAD_PLUGIN_", the suffix will then
be used to define the plugin git directory in lib/plugins/.

It also allows to pick which commit, tag or branch we want to use, by
appending it to the git URL, separated by a '|' for now.

If the commit/tag/branch part is ommited, the default is to use the
git's default branch. The repository is synced at every restart, so it's
probably better to specify which tag or commit hash we want, if this is
a plugin with a lot of activity.

The end result is commented in the docker-compose.yml to avoid confusing
people who use docker and don't need plugins.