Make docker more configurable
At the moment, the docker deployment needs a complete file in a fixed
path (./config/config.js). There was a variable `CPAD_CONF` but it is
not used anymore in the code, I guess it became `CRYPTPAD_CONFIG` but
the docker manifest was never updated.

I think it would be great to get a new interface to configure the
container. For example by passing variables to the container instead of
a big file. I could test and set the environment variable to what looked
the more like what is in config.example.js and made some adjustments
because it is a container.

Changing this way of configuration could break some installation. So I
made sure people using the old fashion would still be able to use their
configuration as a file, but with a warning. Maybe it is a bit brutal to
do this.

It also gives a partial fix for #509 .
The remaining volumes mount path are not configurable atm.

Made this patch for my use case and figured you might want it too. Can fix some
stuff if needed.